V3_Alpha
Contains the convergence behavior of the solutions (Correctness needs to be verified). The plot is not a loglog plot (due to unconfirmed correctness, but change to loglog is very simple). Please check and update the code if there are any inconsistencies. The figure plot of the exact-to-numerical solution is string commented out for focus on the convergence behavior plot (decomment after checking and confirmation).